HU014 - Visualizar un detalle de cada cita

+++ database/migrations/2023_11_01_170803_create_tipocitas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tipocitas', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nombre', 250);
-            $table->integer('tiempo');
+            $table->double('tiempo', 8, 2);
             $table->double('precio', 8, 2);
             $table->timestamps();
         });
